tests/test_simple_anomaly_integration.py
"""
Simple integration tests for Anomaly Detection Service integration
Tests the basic integration functionality
"""

from datetime import datetime
from unittest.mock import Mock, patch
from uuid import uuid4

import pytest

from app.services.anomaly_detection_service import AnomalyDetectionService
from app.services.enhanced_audit_service import EnhancedAuditService


class TestSimpleAnomalyIntegration:
    """Test basic anomaly detection integration"""

    @pytest.fixture
    def mock_db(self):
        """Mock database session"""
        return Mock()

    def test_anomaly_service_basic_functionality(self, mock_db):
        """Test that anomaly service can be instantiated and basic methods exist"""

        # Create anomaly service
        anomaly_service = AnomalyDetectionService(mock_db)

        # Verify service has required methods
        assert hasattr(anomaly_service, "detect_anomalies")
        assert hasattr(anomaly_service, "get_anomaly_summary")
        assert hasattr(anomaly_service, "analyze_trends")
        assert hasattr(anomaly_service, "compare_with_industry_benchmarks")

        # Verify service is properly initialized
        assert anomaly_service.db == mock_db

    def test_audit_service_anomaly_integration_exists(self, mock_db):
        """Test that audit service has anomaly integration code"""

        # Read the audit service file to verify integration exists
        with open("app/services/enhanced_audit_service.py", "r", encoding="utf-8") as f:
            content = f.read()

        # Verify anomaly detection import exists
        assert (
            "from app.services.anomaly_detection_service import AnomalyDetectionService"
            in content
        )

        # Verify anomaly detection is used in create_audit_session
        assert "anomaly_service = AnomalyDetectionService(self.db)" in content
        assert "anomaly_findings" in content
        assert "anomaly_detection_enabled" in content

    def test_validation_service_anomaly_integration_exists(self, mock_db):
        """Test that validation service has anomaly integration code"""

        # Read the validation service file to verify integration exists
        with open(
            "app/services/emissions_validation_service.py", "r", encoding="utf-8"
        ) as f:
            content = f.read()

        # Verify anomaly detection import exists
        assert (
            "from app.services.anomaly_detection_service import AnomalyDetectionService"
            in content
        )

        # Verify anomaly detection is used in validation
        assert "anomaly_service = AnomalyDetectionService(self.db)" in content
        assert "anomaly_report = anomaly_service.detect_anomalies" in content
        assert "_perform_anomaly_detection" in content

    def test_audit_endpoints_anomaly_integration_exists(self):
        """Test that audit endpoints have anomaly integration"""

        # Read the audit endpoints file to verify integration exists
        with open("app/api/v1/endpoints/enhanced_audit.py", "r", encoding="utf-8") as f:
            content = f.read()

        # Verify anomaly detection import exists
        assert (
            "from app.services.anomaly_detection_service import AnomalyDetectionService"
            in content
        )

        # Verify new anomaly endpoints exist
        assert "/companies/{company_id}/anomaly-insights" in content
        assert "/audit-sessions/{session_id}/anomaly-review" in content
        assert "get_audit_anomaly_insights" in content
        assert "create_anomaly_review_task" in content

    def test_integration_error_handling_exists(self):
        """Test that error handling for anomaly integration exists"""

        # Read validation service to check error handling
        with open(
            "app/services/emissions_validation_service.py", "r", encoding="utf-8"
        ) as f:
            validation_content = f.read()

        # Verify error handling exists
        assert "try:" in validation_content
        assert "except Exception as e:" in validation_content
        assert "logger.warning" in validation_content
        assert "Anomaly detection failed during validation" in validation_content

        # Read audit service to check error handling
        with open("app/services/enhanced_audit_service.py", "r", encoding="utf-8") as f:
            audit_content = f.read()

        # Verify error handling exists in audit service too
        assert "try:" in audit_content
        assert "except Exception as e:" in audit_content

    def test_anomaly_schemas_exist(self):
        """Test that anomaly detection schemas are properly defined"""

        from app.schemas.anomaly_detection import (
            AnomalyDetectionRequest,
            AnomalyDetectionResultResponse,
            AnomalyReportResponse,
            AnomalySummaryResponse,
        )

        # Verify schemas can be imported
        assert AnomalyDetectionRequest is not None
        assert AnomalyReportResponse is not None
        assert AnomalyDetectionResultResponse is not None
        assert AnomalySummaryResponse is not None

    def test_anomaly_api_endpoints_exist(self):
        """Test that anomaly API endpoints are properly defined"""

        # Read the anomaly endpoints file
        with open(
            "app/api/v1/endpoints/anomaly_detection.py", "r", encoding="utf-8"
        ) as f:
            content = f.read()

        # Verify main endpoints exist
        assert "/detect" in content
        assert "/summary" in content
        assert "/trends" in content
        assert "/industry-benchmark" in content
        assert "/batch-detect" in content

    def test_integration_documentation_exists(self):
        """Test that integration documentation exists"""

        import os

        # Verify integration guide exists
        assert os.path.exists("ANOMALY_INTEGRATION_GUIDE.md")

        # Read and verify content
        with open("ANOMALY_INTEGRATION_GUIDE.md", "r", encoding="utf-8") as f:
            content = f.read()

        # Verify key sections exist
        assert "Integration Architecture" in content
        assert "Emissions Validation Service Integration" in content
        assert "Enhanced Audit Service Integration" in content
        assert "API Integration" in content
        assert "Error Handling and Resilience" in content

    def test_integration_tests_exist(self):
        """Test that integration test files exist"""

        import os

        # Verify integration test files exist
        assert os.path.exists("tests/test_anomaly_integration.py")
        assert os.path.exists("tests/test_simple_anomaly_integration.py")

        # Verify anomaly detection tests exist
        assert os.path.exists("tests/test_anomaly_detection.py")

    @pytest.mark.asyncio
    async def test_mock_integration_workflow(self, mock_db):
        """Test a mock integration workflow"""

        # Create mock result
        mock_result = Mock()
        mock_result.total_anomalies = 2
        mock_result.overall_risk_score = 75.0
        mock_result.detected_anomalies = []
        mock_result.anomalies_by_severity = {"high": 1, "medium": 1}
        mock_result.summary_insights = ["Test insight"]

        # Mock the detect_anomalies method directly
        with patch.object(
            AnomalyDetectionService, "detect_anomalies", return_value=mock_result
        ) as mock_detect:
            # Test that we can create the service and call methods
            anomaly_service = AnomalyDetectionService(mock_db)

            # Test detection call
            company_id = uuid4()
            user_id = uuid4()

            result = anomaly_service.detect_anomalies(
                company_id=company_id, reporting_year=2024, user_id=user_id
            )

            # Verify the mock was called
            mock_detect.assert_called_once_with(
                company_id=company_id, reporting_year=2024, user_id=user_id
            )

            # Verify result structure
            assert result.total_anomalies == 2
            assert result.overall_risk_score == 75.0
