requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis cache
redis==5.0.1
hiredis==2.2.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client for EPA API
httpx==0.25.2
aiohttp==3.9.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Report generation
reportlab==4.0.7
openpyxl==3.1.2
jinja2==3.1.2

# Environment management
python-dotenv==1.0.0
